big refactoring to eliminte redundant code, cleanup the existing code (renamings), and implementation of ePub merging of XHTML spine. Tested, it works. However, Tobi renderer needs to be updated to understand the XHTML markup (right now it only handles DTBook grammar)

+++ trunk/csharp/XukImport/DaisyToXuk/DaisyToXuk.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.IO;
-using System.Windows.Forms;
+using System.Net;
+using System.Net.Cache;
 using System.Xml;
-using ICSharpCode.SharpZipLib.Zip;
 using urakawa;
-using urakawa.media;
 using urakawa.media.data;
-using urakawa.metadata;
 using urakawa.property.channel;
-using urakawa.property.xml;
 using core = urakawa.core;
 
 namespace XukImport
@@ -16,10 +13,13 @@ namespace XukImport
     public partial class DaisyToXuk
     {
         private readonly string m_outDirectory;
-        private readonly string m_Book_FilePath;
+        private string m_Book_FilePath;
 
         private Project m_Project;
-        private TextChannel m_textChannel;
+        public Project Project
+        {
+            get { return m_Project; }
+        }
 
         public DaisyToXuk(string bookfile, string outDir)
         {
@@ -31,11 +31,10 @@ namespace XukImport
             }
 
             initializeProject();
-
             transformBook();
 
-            Uri uri = new Uri(bookfile + ".xuk");
-            m_Project.SaveXuk(uri);
+            string xukPath = Path.Combine(m_outDirectory, Path.GetFileName(m_Book_FilePath) + ".xuk");
+            m_Project.SaveXuk(new Uri(xukPath));
         }
 
         public DaisyToXuk(string bookfile)
@@ -93,11 +92,11 @@ namespace XukImport
             m_audioChannel = presentation.ChannelFactory.CreateAudioChannel();
             m_audioChannel.Name = "Our Audio Channel";
 
-            string dataPath = presentation.DataProviderManager.DataFileDirectoryFullPath;
-            /*if (Directory.Exists(dataPath))
-            {
-                Directory.Delete(dataPath, true);
-            }*/
+            /*string dataPath = presentation.DataProviderManager.DataFileDirectoryFullPath;
+           if (Directory.Exists(dataPath))
+           {
+               Directory.Delete(dataPath, true);
+           }*/
         }
 
         private void transformBook()
@@ -116,27 +115,24 @@ namespace XukImport
             {
                 case ".opf":
                     {
-                        parseOPFAndPopulateDataModel();
+                        XmlDocument opfXmlDoc = readXmlDocument(m_Book_FilePath);
+                        parseOpf(opfXmlDoc);
                         break;
                     }
                 case ".xml":
                     {
-                        XmlDocument bookXmlDoc = readXmlDocument(m_Book_FilePath);
-                        parseDTBookXmlDocAndPopulateDataModel(bookXmlDoc, null);
+                        XmlDocument contentXmlDoc = readXmlDocument(m_Book_FilePath);
+                        parseContentDocument(contentXmlDoc, null);
                         break;
                     }
                 case ".epub":
                     {
-                        unZipePub();
+                        unzipEPubAndParseOpf();
                         break;
                     }
                 default:
                     break;
             }
-        }
-        public Project Project
-        {
-            get { return m_Project; }
         }
 
         private XmlDocument readXmlDocument(string path)
@@ -144,28 +140,33 @@ namespace XukImport
             XmlReaderSettings settings = new XmlReaderSettings();
 
             settings.ProhibitDtd = false;
-            settings.XmlResolver = null;
+            settings.ValidationType = ValidationType.None;
+            settings.ConformanceLevel = ConformanceLevel.Auto;
+
+            settings.XmlResolver = new LocalXmlUrlResolver(true);
 
             settings.IgnoreComments = true;
             settings.IgnoreProcessingInstructions = true;
             settings.IgnoreWhitespace = true;
 
-            XmlReader xmlReader = XmlReader.Create(path, settings);
-
-            XmlDocument xmldoc = new XmlDocument();
-            xmldoc.XmlResolver = null;
-            try
+            using (XmlReader xmlReader = XmlReader.Create(path, settings))
             {
-                xmldoc.Load(xmlReader);
-            }
-            finally
-            {
-                xmlReader.Close();
-            }
 
-            return xmldoc;
+                XmlDocument xmldoc = new XmlDocument();
+                xmldoc.XmlResolver = null;
+                try
+                {
+                    xmldoc.Load(xmlReader);
+                }
+                finally
+                {
+                    xmlReader.Close();
+                }
+
+                return xmldoc;
+            }
         }
-        
+
         private core.TreeNode getTreeNodeWithXmlElementId(string id)
         {
             Presentation pres = m_Project.GetPresentation(0);
@@ -186,124 +187,53 @@ namespace XukImport
             }
             return null;
         }
+    }
 
-        private void parseDTBookXmlDocAndPopulateDataModel(XmlNode xmlNode, core.TreeNode parentTreeNode)
+    public class LocalXmlUrlResolver : XmlUrlResolver
+    {
+        bool enableHttpCaching;
+        ICredentials credentials;
+
+        //resolve resources from cache (if possible) when enableHttpCaching is set to true
+        //resolve resources from source when enableHttpcaching is set to false 
+        public LocalXmlUrlResolver(bool enableHttpCaching)
         {
-            XmlNodeType xmlType = xmlNode.NodeType;
-            switch (xmlType)
+            this.enableHttpCaching = enableHttpCaching;
+        }
+
+        public override ICredentials Credentials
+        {
+            set
             {
-                case XmlNodeType.Attribute:
-                    {
-                        System.Diagnostics.Debug.Fail("Calling this method with an XmlAttribute should never happen !!");
-                        break;
-                    }
-                case XmlNodeType.Document:
-                    {
-                        parseDTBookXmlDocAndPopulateDataModel(((XmlDocument)xmlNode).DocumentElement, parentTreeNode);
-                        break;
-                    }
-                case XmlNodeType.Element:
-                    {
-                        Presentation presentation = m_Project.GetPresentation(0);
-
-                        core.TreeNode treeNode = presentation.TreeNodeFactory.Create();
-
-                        if (parentTreeNode == null)
-                        {
-                            presentation.RootNode = treeNode;
-                            parentTreeNode = presentation.RootNode;
-                        }
-                        else
-                        {
-                            parentTreeNode.AppendChild(treeNode);
-                        }
-
-                        XmlProperty xmlProp = presentation.PropertyFactory.CreateXmlProperty();
-                        treeNode.AddProperty(xmlProp);
-                        xmlProp.LocalName = xmlNode.Name;
-                        if (xmlNode.ParentNode != null && xmlNode.ParentNode.NodeType == XmlNodeType.Document)
-                        {
-                            presentation.PropertyFactory.DefaultXmlNamespaceUri = xmlNode.NamespaceURI;
-                        }
-
-                        if (xmlNode.NamespaceURI != presentation.PropertyFactory.DefaultXmlNamespaceUri)
-                        {
-                            xmlProp.NamespaceUri = xmlNode.NamespaceURI;
-                        }
-
-                        XmlAttributeCollection attributeCol = xmlNode.Attributes;
-
-                        if (attributeCol != null)
-                        {
-                            for (int i = 0; i < attributeCol.Count; i++)
-                            {
-                                XmlNode attr = attributeCol.Item(i);
-                                if (attr.Name != "smilref")
-                                {
-                                    xmlProp.SetAttribute(attr.Name, "", attr.Value);
-                                }
-                            }
-
-
-                            if (xmlNode.Name == "meta")
-                            {
-                                XmlNode attrName = attributeCol.GetNamedItem("name");
-                                XmlNode attrContent = attributeCol.GetNamedItem("content");
-                                if (attrName != null && attrContent != null && !String.IsNullOrEmpty(attrName.Value)
-                                        && !String.IsNullOrEmpty(attrContent.Value))
-                                {
-                                    Metadata md = presentation.MetadataFactory.CreateMetadata();
-                                    md.Name = attrName.Value;
-                                    md.Content = attrContent.Value;
-                                    presentation.AddMetadata(md);
-                                }
-                            }
-                        }
-
-                        foreach (XmlNode childXmlNode in xmlNode.ChildNodes)
-                        {
-                            parseDTBookXmlDocAndPopulateDataModel(childXmlNode, treeNode);
-                        }
-                        break;
-                    }
-                case XmlNodeType.Text:
-                    {
-                        Presentation presentation = m_Project.GetPresentation(0);
-
-                        string text = xmlNode.Value;
-                        TextMedia textMedia = presentation.MediaFactory.CreateTextMedia();
-                        textMedia.Text = text;
-
-                        ChannelsProperty cProp = presentation.PropertyFactory.CreateChannelsProperty();
-                        cProp.SetMedia(m_textChannel, textMedia);
-
-                        int counter = 0;
-                        foreach (XmlNode childXmlNode in xmlNode.ParentNode.ChildNodes)
-                        {
-                            XmlNodeType childXmlType = childXmlNode.NodeType;
-                            if (childXmlType == XmlNodeType.Text || childXmlType == XmlNodeType.Element)
-                            {
-                                counter++;
-                            }
-                        }
-                        if (counter == 1)
-                        {
-                            parentTreeNode.AddProperty(cProp);
-                        }
-                        else
-                        {
-                            core.TreeNode txtWrapperNode = presentation.TreeNodeFactory.Create();
-                            txtWrapperNode.AddProperty(cProp);
-                            parentTreeNode.AppendChild(txtWrapperNode);
-                        }
-
-                        break;
-                    }
-                default:
-                    {
-                        return;
-                    }
+                credentials = value;
+                base.Credentials = value;
             }
         }
-    }//class
-}//namespace
+
+        public override object GetEntity(Uri absoluteUri, string role, Type ofObjectToReturn)
+        {
+            if (absoluteUri == null)
+            {
+                throw new ArgumentNullException("absoluteUri");
+            }
+            //resolve resources from cache (if possible)
+            if (absoluteUri.Scheme == "http" && enableHttpCaching && (ofObjectToReturn == null || ofObjectToReturn == typeof(Stream)))
+            {
+                WebRequest webReq = WebRequest.Create(absoluteUri);
+                webReq.CachePolicy = new HttpRequestCachePolicy(HttpRequestCacheLevel.Default);
+                if (credentials != null)
+                {
+                    webReq.Credentials = credentials;
+                }
+                WebResponse resp = webReq.GetResponse();
+                return resp.GetResponseStream();
+            }
+            //otherwise use the default behavior of the XmlUrlResolver class (resolve resources from source)
+            else
+            {
+                return base.GetEntity(absoluteUri, role, ofObjectToReturn);
+            }
+        }
+
+    }
+}
